+++ src/omni_epd/virtualepd.py
@@ -169,7 +169,11 @@ class VirtualEPD:
             palette_image = Image.new("P", (1, 1))
 
             # set the palette, set all other colors to 0
-            palette_image.putpalette(palette + [0, 0, 0] * (256-len(palette)))
+            palette_image.putpalette(palette + [0, 0, 0] * (256-len(colors)))
+
+            if(image.mode != 'RGB'):
+                # convert to RGB as quantize requires it
+                image = image.convert(mode='RGB')
 
             # apply the palette
             image = image.quantize(palette=palette_image, dither=dither)