+++ src/omni_epd/displays/waveshare_display.py
@@ -215,13 +215,13 @@ class WaveshareTriColorDisplay(WaveshareDisplay):
             # separate out black from the image
             img_black = image.copy()
 
-            # convert greys to black or white based on threshold (default value: 16)
+            # convert greys to black or white based on threshold of 20
             img_black = img_black.point(lambda p: 255 if p >= 20 else 0)
 
-            # separate out red from the image
+            # separate out third color from the image
             img_color = image.copy()
 
-            # convert greys to red (represented as black in the image) or white based on threshold (default value: 16)
+            # convert greys to third color (represented as black in the image) or white based on threshold of 20
             img_color = img_color.point(lambda p: 0 if 20 < p < 235 else 255)
 
             # send to display