+++ tests/test_image_processing.py
@@ -4,10 +4,12 @@ import time
 import glob
 import pytest
 from . import constants as constants
-from PIL import Image
+from PIL import Image, ImageChops
 from shutil import copyfile
 from omni_epd import displayfactory
 from omni_epd.conf import CONFIG_FILE
+
+TEST_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__)))
 
 
 class TestImageProcessing(unittest.TestCase):
@@ -23,25 +25,96 @@ class TestImageProcessing(unittest.TestCase):
     def run_before_and_after_tests(self):
         # clean up any files left over from previous tests
         self._delete_files()
-        self._delete_files('.png')
+        self._delete_files('png')
         yield
 
         # clean up any files made during this test
         self._delete_files()
-        self._delete_files('.png')
+        self._delete_files('png')
+
+    def setup_config(self, source_config_file_name, target_config_file_name):
+        copyfile(os.path.join(TEST_PATH, 'ini', source_config_file_name), os.path.join(os.getcwd(), target_config_file_name))
+        time.sleep(1)
+
+    def open_image(self, image, w, h):
+        """Open an image and resize it for EPD display"""
+        result = Image.open(image)
+
+        return result.resize((w, h))
+
+    def compare_images(self, image_one, image_two):
+        """compare if two images are equal, return true/false """
+        im1 = Image.open(image_one)
+        im2 = Image.open(image_two)
+
+        diff = ImageChops.difference(im1, im2)
+
+        if diff.getbbox() is None:
+            # same
+            return True
+        else:
+            return False
 
     def test_image_processing_options(self):
         """
-        Test all common image processing options (rotating, dithering, contrast, etc)
+        Test all common image processing options (rotating, contrast, etc)
         https://github.com/robweber/omni-epd#advanced-epd-control
         """
 
-        copyfile(os.path.join(os.getcwd(), "tests", constants.ALL_IMAGE_OPTIONS), os.path.join(os.getcwd(), CONFIG_FILE))
-        time.sleep(1)
+        self.setup_config(constants.ALL_IMAGE_OPTIONS, CONFIG_FILE)
 
         epd = displayfactory.load_display_driver(constants.GOOD_EPD_NAME)
 
         # write the image
-        image = Image.open(constants.GALAXY_IMAGE)
+        image = self.open_image(constants.GALAXY_IMAGE, epd.width, epd.height)
 
         epd.display(image)
+
+    def test_basic_dither(self):
+        """
+        Test that a basic dither algorithm can be applied - tests that result image is different than master (non-modified) image
+        Dithering will return same image if not applied or dither algorithm does not exist
+        """
+        self.setup_config(constants.BASIC_DITHER, CONFIG_FILE)
+
+        epd = displayfactory.load_display_driver(constants.GOOD_EPD_NAME)
+
+        # write the image
+        image = self.open_image(constants.GALAXY_IMAGE, epd.width, epd.height)
+        epd.display(image)
+
+        # compare the two images should be different (dither applied)
+        assert not self.compare_images(constants.MOCK_EPD_OUTPUT, constants.MASTER_IMAGE)
+
+    def test_custom_dither(self):
+        """
+        Tests that custom dithering can be applied via either the INI file
+        Tests that generated images are not the same as a master (non-modified image)
+        """
+        self.setup_config(constants.CUSTOM_DITHER_INI, CONFIG_FILE)
+
+        epd = displayfactory.load_display_driver(constants.GOOD_EPD_NAME)
+
+        # write the image
+        image = self.open_image(constants.GALAXY_IMAGE, epd.width, epd.height)
+        epd.display(image)
+
+        # compare the two images should be different (dither applied)
+        assert not self.compare_images(constants.MOCK_EPD_OUTPUT, constants.MASTER_IMAGE)
+
+    def test_custom_dither_json(self):
+        """
+        Tests that custom dithering can be applied from a JSON file
+        Tests that generated images are not the same as a master (non-modified image)
+        This will fail if JSON file can't be loaded, the same image will be returned by didder
+        """
+        self.setup_config(constants.CUSTOM_DITHER_JSON, CONFIG_FILE)
+
+        epd = displayfactory.load_display_driver(constants.GOOD_EPD_NAME)
+
+        # write the image
+        image = self.open_image(constants.GALAXY_IMAGE, epd.width, epd.height)
+        epd.display(image)
+
+        # compare the two images should be different (dither applied)
+        assert not self.compare_images(constants.MOCK_EPD_OUTPUT, constants.MASTER_IMAGE)