+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - calls to `Image.quantize` require an RGB or L mode Image object, convert any loaded image before applying new palettes
 - when filling palette too many colors were being set (< 256), wrong length variable was being used
 
+### Changed
+
+- make sure Pillow and Inky packages are known working versions or above - thanks @donbing
+
 ## 0.2.6
 
 ### Added