+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## Unreleased
+
+### Fixed
+
+- calls to `Image.quantize` require an RGB or L mode Image object, convert any loaded image before applying new palettes
+- when filling palette too many colors were being set (< 256), wrong length variable was being used
+
+### Changed
+
+- make sure Pillow and Inky packages are known working versions or above - thanks @donbing
+- `bw` standardized as the consistent naming for the default black/white device mode. `black` will throw a warning, affects Inky devices - thanks @missionfloyd
+
 ## 0.2.6
 
 ### Added