+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - calls to `Image.quantize` require an RGB or L mode Image object, convert any loaded image before applying new palettes
 - when filling palette too many colors were being set (< 256), wrong length variable was being used
 - fixed regression where Inky `bw` mode was causing colors to be inverted
+- universal fix for Waveshare Tri-color displays as original fixes broke some displays - thanks @aaron8684
 
 ### Changed