+++ CHANGELOG.md
@@ -4,11 +4,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## Unreleased
+## Version 0.3.1
 
 ### Added
 
 - `omni_epd.mock` can now set both the width and height values within the `.ini` file. Thanks @missionfloyd
+
+### Fixed
+
+- `omni_epd.mock` device now returns a palette filter when using the color mode. Previously this returned only b/w and resulted in image processing enhancements resulting in a black and white only image, even when color was selected. Hardcoded palette based on web safe colors.
+
+- Waveshare device `epd2in13_V2` should use the alternate clear method which requires a color parameter. Thanks @ThatIsAPseudo for pointing this out
+
+### Changed
+
+- dithering is now done with [didder](https://github.com/makeworld-the-better-one/didder) - this is a massive improvement both in scope and speed to hitherdither. Thanks @missionfloyd
+
+### Removed
+
+- removed `hitherdither` as a dependency
 
 ## 0.3.0